+++ crates/resvg/examples/custom_href_resolver.rs
@@ -1,5 +1,3 @@
-use usvg::TreeParsing;
-
 fn main() {
     let mut opt = usvg::Options::default();