+++ crates/resvg/examples/minimal.rs
@@ -1,4 +1,4 @@
-use usvg::{fontdb, TreeParsing, TreePostProc};
+use usvg::fontdb;
 
 fn main() {
     let args: Vec<String> = std::env::args().collect();